refactor: alterado tipagem para BigDecimal.

+++ food-backend/src/main/java/com/softexpert/food/domain/model/ItemPedido.java
@@ -4,6 +4,7 @@ import lombok.Builder;
 import lombok.Data;
 
 import javax.validation.constraints.PositiveOrZero;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -12,5 +13,5 @@ public class ItemPedido {
     private String tituloItem;
 
     @PositiveOrZero(message = "O valor do item não pode ser negativo.")
-    private double valorItem;
+    private BigDecimal valorItem;
 }
