+++ food-backend/src/test/java/com/softexpert/food/domain/model/PedidoTest.java
@@ -3,6 +3,7 @@ package com.softexpert.food.domain.model;
 import org.junit.jupiter.api.Test;
 
 import javax.validation.Valid;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -31,7 +32,7 @@ public class PedidoTest {
     @Test
     public void testToString() {
         List<ItemPedido> itensPedido = new ArrayList<>();
-        itensPedido.add(new ItemPedido("Livro", 20.0));
+        itensPedido.add(new ItemPedido("Livro", BigDecimal.valueOf(20.0)));
         Pedido pedido = new Pedido(itensPedido, "Carlos");
 
         assertThat(pedido.toString()).isEqualTo("Pedido(itensPedido=[ItemPedido(tituloItem=Livro, valorItem=20.0)], nomeSolicitante=Carlos)");
@@ -40,13 +41,13 @@ public class PedidoTest {
     @Test
     public void testEquals() {
         List<ItemPedido> itensPedido1 = new ArrayList<>();
-        itensPedido1.add(new ItemPedido("Livro", 20.0));
+        itensPedido1.add(new ItemPedido("Livro", BigDecimal.valueOf(20.0)));
 
         List<ItemPedido> itensPedido2 = new ArrayList<>();
-        itensPedido2.add(new ItemPedido("Livro", 20.0));
+        itensPedido2.add(new ItemPedido("Livro", BigDecimal.valueOf(20.0)));
 
         List<ItemPedido> itensPedido3 = new ArrayList<>();
-        itensPedido3.add(new ItemPedido("Caneta", 5.0));
+        itensPedido3.add(new ItemPedido("Caneta", BigDecimal.valueOf(5.0)));
 
         Pedido pedido1 = new Pedido(itensPedido1, "Ana");
         Pedido pedido2 = new Pedido(itensPedido2, "Ana");
@@ -59,13 +60,13 @@ public class PedidoTest {
     @Test
     public void testHashCode() {
         List<ItemPedido> itensPedido1 = new ArrayList<>();
-        itensPedido1.add(new ItemPedido("Livro", 20.0));
+        itensPedido1.add(new ItemPedido("Livro", BigDecimal.valueOf(20.0)));
 
         List<ItemPedido> itensPedido2 = new ArrayList<>();
-        itensPedido2.add(new ItemPedido("Livro", 20.0));
+        itensPedido2.add(new ItemPedido("Livro", BigDecimal.valueOf(20.0)));
 
         List<ItemPedido> itensPedido3 = new ArrayList<>();
-        itensPedido3.add(new ItemPedido("Caneta", 5.0));
+        itensPedido3.add(new ItemPedido("Caneta", BigDecimal.valueOf(5.0)));
 
         Pedido pedido1 = new Pedido(itensPedido1, "Ana");
         Pedido pedido2 = new Pedido(itensPedido2, "Ana");
